requirements.txt
openai-whisper
requests
